requirements.txt
git+https://github.com/huggingface/diffusers
transformers>=4.49.0
accelerate>=1.2.1
safetensors>=0.4.0
pillow>=10.0.0
runpod
bitsandbytes>=0.45.0
peft>=0.14.0
torchvision>=0.20.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
# Queue worker mode dependencies
redis>=5.0.0
boto3>=1.34.0

# Testing
pytest>=8.0.0
pytest-mock>=3.12.0
